Add support for save and load feature
Chatbot does not save tasks in the hard disk. It does not retrieve
tasks from hard disk.

Chatbot should be able to save and load tasks so that users can
access them when the chatbot is restarted.

Write tasks into a text file after each update to the tasklist. When
the chatbot starts up, retrieve tasks from file.

Saving tasks into a text file and retrieving tasks from a text file
is simple. The tasks are readable directly from the file.