Add support for reading dates and times
Chatbot is not able to understand strings in date and time format.

It is easier to understand date and time format

Add functionality to read date and time in the yyyy-MM-dd HHmm format.
Add functionality to print date and time in the MMM d yyyy hh:mm am/pm
format

These formats are easy to use and read