Add GUI using JavaFX
Chatbot is CLI based.

CLI based chatbot is less visually appealing that GUI. Having a GUI
makes it more accessible.

Implement GUI using JavaFX. The commands remain command line based.

JavaFX is used to create GUIs in Java. The commands are command
line based as it is faster to type everything in a single line than
filling in multiple blanks.