Add assertions
The code does not have any assertions.

There is no way to check if functions that do not return null
encounter a situation where they return null. Additionally,
there is no way to ensure the code does not go to areas it
should not go to.

Let's
* Assert that objects returned from functions are not null
* Assert that code does not reach outside of switch statements if
it does not have to

Using assertions will allow for early detection of bugs.